tweaks

+++ measurements/experiments-aws-single/analyze_rsl.py
@@ -178,8 +178,8 @@ def analyze_client_csv(filepath):
         for row in csvreader:
             if 'TIMEOUT' in row[0] or 'DEBUG:' in row:
                 continue
-            req_start = int(row[1])
-            req_end = int(row[2])
+            req_start = float(row[1])
+            req_end = float(row[2])
             start = min(start, req_start)   # Note: this is rather inefficient
             end = max(end, req_end)
             durations_milli.append(req_end - req_start)
